fix(boson_sampling_utilities): Lexicographic order (#13)
**Problem**

The Fock basis states were not ordered. In the calculations the ordering
does not matter provided that the same ordering is used everywhere, but
`BSDistributionCalculatorWithFixedLosses.calculate_distribution()`
exposes this ordering. This order has to be fixed.

**Solution**

It would be sensible to perform the ordering firstly by the number of
particles. Among the Fock states corresponding to the same number of
particles, the order could be lexicographic, i.e. for 2 modes and up to
3 particles one would get
```
00, 01, 10, 02, 11, 20, 03, 12, 21, 30.
```

Ordering by the number of particles is already present in the code.

The `generate_possible_n_particle_outputs` produces all the possible
Fock states corresponding to a fixed number of particles, but the result
was not sorted.

To lexicographically sort Fock states, one could use `numpy.lexsort`,
but its usage is quite hairy: it would require to concatenate all the
Fock states together to sort them. Instead, the Fock states got
converted into tuple, which could be sorted easily.

A test case has been written to test the value of the exact
distribution.

Co-authored-by: Zoltán Kolarovszki <zk@localhost.localdomain>